==> 02_1_TF_bs_similarity_dimension_reduction_Arabidopsis_heatmaps.Rmd <==
---
title: "Heatmaps: Analysis of TF binding site similarity in Arabidopsis"
output: html_notebook
---

Aim: To compare predicted binding sites of TFs and find % of binding sites of different TFs overlaping by 4 or more base pairs
data_input: Arabidopsis: analysed_data/TF_rc_scan_results_promoters1500_TF_family.csv analysed in 02_TF_bs_similarity_dimension_reducion_Arabidopsis
            
            

# load packages:
```{r}
source("scripts/pkgs.R")
```

##--Load the data containing overlaps calculated from 02_TF_bs_similarity_dimension_reducion_Arabidopsis
```{r}
Results_overlap_TFs <- read_csv("../TF_binding/output_data/02_TF_bs_similarity_dimension_reduction_Arabidopsis/Results_overlap_TFs_all.csv")

```
pct_overlap_TF1 = round(overlap_TF_1/tot_TF1, 2)
pct_overlap_TF2 = round(overlap_TF_2/tot_TF2, 2)

# load data = the results of predicted TF's binding sites in 1500 bp promoters in Arabidopsis
```{r}
TF_scan <- read_csv("../TF_binding/analysed_data/TF_rc_scan_results_promoters1500_TF_family.csv")

```

# Calculate total number of binding site for each TF; order as TF_list
```{r}

N_TF_bs <- TF_scan %>% 
  count(motif)

```

##--Hearmap

Build matrix of overlaps between all TFs, the digonal will be 100%

```{r}
overlap_TFs_matrix <- bind_rows(
# part_1
 (Results_overlap_TFs %>% 
   select(motif_1, motif_2, overlap = overlap_TF_1, tot = tot_TF1)),
   
# part_2 
(Results_overlap_TFs %>% 
   select(motif_2 = motif_1, motif_1 = motif_2, overlap = overlap_TF_2, tot = tot_TF2)),
# part_3
 (N_TF_bs %>%
  select(motif_1 = motif, tot = n) %>%
  mutate(motif_2 = motif_1, overlap = tot))
 ) %>%

  mutate(pct_overlap = round(overlap/tot*100,3)) %>%
  select(-overlap, -tot) %>%
  spread(motif_2, pct_overlap)


which(is.na(overlap_TFs_matrix), arr.ind = TRUE)

```

##--Clustering 
###--Define parameters:
```{r}
source("~/Work/2019/Analysis/Utilities_R/Functions/My_functions.R")

data_matrix <- overlap_TFs_matrix
Title_plot <- "hclust_TFs_overlaps" # main= Title for the plot; string
Path_to_output_results <- "../TF_binding/output_data/"
Path_to_plot <- "../TF_binding/plots/"

```


#### Clustering TF binding site by hclust.person / columns
```{r eval=FALSE}
#arguments:
hclust_data <- data_matrix %>% column_to_rownames(.,var = "motif_1") #Input matrix to cluster

k_row = 20  # n of clusters in cutree 
k_col = 20


my_brewer_par <- c(brewer.pal(12,'Paired'),brewer.pal(8,'Set1'))


hclust_rows <- hclust.person.rows(hclust_data, k_row) #from my_functions
names(hclust_rows)
#[1] "Rowv_arg" "RowSideColors_arg" "Cluster_rows"   

hclust_columns <- hclust.person.columns(hclust_data, k_col)
names(hclust_columns)
#[1] "Colv_arg" "ColSideColors_arg" "Cluster_columns"  

```


##-- Create heatmap
###-- Define parametrs
```{r eval=FALSE}
#arguments:
data_plot <- hclust_data #Input matrix to plot

column_names <- str_remove(colnames(data_plot), "\\..*$") #labCol=column_names; vector of the names to be ploted for each column in heatmap

#only for dendograms add the folowing arguments to create sep horizontal lines in heatmap; #line/break after columns/row number in vector
## rows

row_sep <- table(hclust_rows$Cluster_rows)

## columns
ind_c <- match(labels(hclust_columns$Colv_arg), colnames(hclust_data))
col_sep <- table(hclust_columns$Cluster_columns)[match(unique(hclust_columns$Cluster_columns[ind_c]),
           names(table(hclust_columns$Cluster_columns)))]

rowsep_arg <- sep_calc(row_sep)
colsep_arg <- sep_calc(col_sep)


########
length(hclust_rows$RowSideColors_arg)
dim(data_plot)

table(hclust_rows$RowSideColors_arg)

```

##-- Plot heatmap dendogram rows and columns 
```{r eval=FALSE}
pdf(paste0(Path_to_plot, Title_plot,"_Rowv_Colv_TF", ".pdf"))
par(cex.main=1) #controls the size of the title
data_heatmap_both <- 
                heatmap.2(as.matrix(data_plot),
                          Rowv = hclust_rows$Rowv_arg,
                          #Rowv = FALSE,
                          Colv = hclust_columns$Colv_arg,
                          dendrogram = 'column',
                          #dendrogram = 'both',
                          #scale = "column",
                          col = colorRampPalette(c("white", "red"))(n = 100),
                          breaks=seq(0, 1,length.out=101), 
                          trace = 'none',         
                          labRow=NA,  
                          labCol=NA,
                          main=Title_plot,
                          RowSideColors = hclust_rows$RowSideColors_arg, 
                          ColSideColors = hclust_columns$ColSideColors_arg, 
                          key=TRUE,
                          #density.info = "none",  #remove the hist from key
                          #lhei = c(2, 8), # this makes the colour-key legend a little thinner,
                          #sepwidth=c(0.05,0.05),
                          #sepcolor="white",
                          # colsep = colsep_arg,
                          # rowsep = rowsep_arg,
                          # the margins command sets the width of the white space around the plot. 
                          #The first element is the bottom margin and the second is the right margin
                          margins = c(12, 2)
)
dev.off()


jpeg(file = paste0(Path_to_plot, Title_plot,"_Rowv_Colv_TF", ".jpeg"))
     heatmap.2(as.matrix(data_plot),
                          Rowv = hclust_rows$Rowv_arg,
                          #Rowv = FALSE,
                          Colv = hclust_columns$Colv_arg,
                          dendrogram = 'column',
                          #dendrogram = 'both',
                          #scale = "column",
                          col = colorRampPalette(c("white", "red"))(n = 100),
                          breaks=seq(0, 1,length.out=101), 
                          trace = 'none',         
                          labRow=NA,  
                          labCol=NA,
                          main=Title_plot,
                          RowSideColors = hclust_rows$RowSideColors_arg, 
                          ColSideColors = hclust_columns$ColSideColors_arg, 
                          key=TRUE,
                          #density.info = "none",  #remove the hist from key
                          #lhei = c(2, 8), # this makes the colour-key legend a little thinner,
                          #sepwidth=c(0.05,0.05),
                          #sepcolor="white",
                          # colsep = colsep_arg,
                          # rowsep = rowsep_arg,
                          # the margins command sets the width of the white space around the plot. 
                          #The first element is the bottom margin and the second is the right margin
                          margins = c(12, 2)
)
dev.off() 


data_heatmap_rows <- 
 heatmap.2(as.matrix(data_plot),
                          Rowv = hclust_rows$Rowv_arg,
                          #Rowv = FALSE,
                          #Colv = hclust_columns$Colv_arg,
                          Colv = FALSE,
                          dendrogram = 'column',
                          #dendrogram = 'both',
                          #scale = "column",
                          col = colorRampPalette(c("white", "red"))(n = 100),
                          breaks=seq(0, 1,length.out=101), 
                          trace = 'none',         
                          labRow=NA,  
                          labCol=NA,
                          main=Title_plot,
                          RowSideColors = hclust_rows$RowSideColors_arg, 
                          ColSideColors = hclust_columns$ColSideColors_arg, 
                          key=TRUE,
                          #density.info = "none",  #remove the hist from key
                          #lhei = c(2, 8), # this makes the colour-key legend a little thinner,
                          #sepwidth=c(0.05,0.05),
                          #sepcolor="white",
                          # colsep = colsep_arg,
                          # rowsep = rowsep_arg,
                          # the margins command sets the width of the white space around the plot. 
                          #The first element is the bottom margin and the second is the right margin
                          margins = c(12, 2)
)

 heatmap.2(as.matrix(data_plot),
                          #Rowv = hclust_rows$Rowv_arg,
                          Rowv = FALSE,
                          Colv = hclust_columns$Colv_arg,
                          #Colv = FALSE,
                          dendrogram = 'column',
                          #dendrogram = 'both',
                          #scale = "column",
                          col = colorRampPalette(c("white", "red"))(n = 100),
                          breaks=seq(0, 1,length.out=101), 
                          trace = 'none',         
                          labRow=NA,  
                          labCol=NA,
                          main=Title_plot,
                          RowSideColors = hclust_rows$RowSideColors_arg, 
                          ColSideColors = hclust_columns$ColSideColors_arg, 
                          key=TRUE,
                          #density.info = "none",  #remove the hist from key
                          #lhei = c(2, 8), # this makes the colour-key legend a little thinner,
                          #sepwidth=c(0.05,0.05),
                          #sepcolor="white",
                          # colsep = colsep_arg,
                          # rowsep = rowsep_arg,
                          # the margins command sets the width of the white space around the plot. 
                          #The first element is the bottom margin and the second is the right margin
                          margins = c(12, 2)
)

```


###--insert Cluster number to data_matrix and reorder
```{r}
data_matrix_cluster_reordered <- data_matrix %>% 
  mutate(cluster_rows = hclust_rows$Cluster_rows) %>%
  dplyr::slice(data_heatmap_both$rowInd) %>% 
   bind_rows( hclust_columns$Cluster_columns[data_heatmap_both$colInd], .) 


data_matrix_cluster_reordered[1:10, 870:874]
# data_matrix_cluster_reordered %>% select((data_heatmap_both$colInd+1)[1:2])
# data_plot %>% select((data_heatmap_both$colInd)[1:2])


dim(data_matrix_cluster_reordered$motif_1)

write_csv(data_matrix_cluster_reordered, "output_data/02_TF_bs_similarity_dimension_reduction_Arabidopsis/Results_overlap_TFs_all_cluster_reordered.csv")

```


##-- Plot heatmap for differnt thresholds for % of overlaps:

```{r}

#Function to plot heatmaps of data after converting probability of overlap X to be 100%
heatmap_filtered_data <- function(overlap_par, data, Path_to_plot, Title_plot, Rowv_arg, Colv_arg, RowSideColors_arg, ColSideColors_arg ){
  
  data_plot <- hclust_data %>% 
  mutate_all(.,funs(if_else(.> overlap_par, 100, 0)))

jpeg(paste0(Path_to_plot, Title_plot,"_Rowv_Colv_TF", "_overlap_", as.character(overlap_par), ".jpeg"))
par(cex.main=1) #controls the size of the title

                heatmap.2(as.matrix(data_plot),
                          Rowv = Rowv_arg,
                          #Rowv = FALSE,
                          Colv = Colv_arg,
                          dendrogram = 'column',
                          #dendrogram = 'both',
                          #scale = "column",
                          col = colorRampPalette(c( "white", "red"))(n = 100),
                          breaks=seq(0, 1,length.out=101),  
                          trace = 'none',         
                          labRow=NA,  
                          labCol=NA,
                          main=paste(Title_plot,"_overlap_", as.character(overlap_par)),
                          RowSideColors = RowSideColors_arg, 
                          ColSideColors = ColSideColors_arg, 
                          key=TRUE,
                          #density.info = "none",  #remove the hist from key
                          #lhei = c(2, 8), # this makes the colour-key legend a little thinner,
                          #sepwidth=c(0.05,0.05),
                          #sepcolor="white",
                          # colsep = colsep_arg,
                          # rowsep = rowsep_arg,
                          # the margins command sets the width of the white space around the plot. 
                          #The first element is the bottom margin and the second is the right margin
                          margins = c(12, 2)
)

dev.off()
}


#Parameters:
overlap_list <- c(12.5, 25, 50, 75, 90)
Colv <- hclust_columns$Colv_arg 
Rowv <-   hclust_rows$Rowv_arg
RowSideColors <- hclust_rows$RowSideColors_arg
ColSideColors <- hclust_columns$ColSideColors_arg

lapply(overlap_list, function(x)heatmap_filtered_data(overlap_par = x, data = hclust_data, Path_to_plot, Title_plot, Rowv_arg = Rowv, Colv_arg = Colv, RowSideColors_arg = RowSideColors, ColSideColors_arg = ColSideColors))
       
```



```{r}

```



```{r}

Clust_row <- as.data.frame(hclust_rows$Cluster_rows) %>% rownames_to_column(., var = "motif")
Clust_col <- as.data.frame(hclust_columns$Cluster_columns) %>% rownames_to_column(., var = "motif")
Clust <- full_join(Clust_row,Clust_col) %>% 
  select(motif, Cluster_rows = `hclust_rows$Cluster_rows`, Cluster_columns = `hclust_columns$Cluster_columns`) %>%
  mutate(TF_family = str_remove(motif, "\\..*$"))

Clust %>% 
group_by(Cluster_rows, Cluster_columns) %>% count(TF_family)

test <- Clust %>% 
group_by(Cluster_rows) %>% count(Cluster_columns) %>% arrange(Cluster_rows, n)

```

```{r}
Clust %>% filter(Cluster_rows == 16) 
  group_by(Cluster_columns) %>% count()

Clust %>% filter(Cluster_rows == 16, Cluster_columns == 2) %>% 
  group_by(TF_family) %>% count()

Clust %>% filter(Cluster_rows == 8, Cluster_columns == 5) %>% 
  group_by(TF_family) %>% count()

Clust %>% filter(Cluster_rows == 7, Cluster_columns == 6) %>% 
  group_by(TF_family) %>% count()

Clust %>% filter(Cluster_rows == 19, Cluster_columns == 17) %>% 
  group_by(TF_family) %>% count()

Clust %>% filter(Cluster_rows == 10) %>% 
  group_by(TF_family) %>% count()


```

clust_colmn	cluster_rows
2	16
5	8
6	7
4	19
16	10
12	9
17	1
9	5
7	12
19	17
15	18
18	2
8	15
3	3
14	11
10	6
11	14
20	20
13	4
1	13

##--top_TF_rf 
```{r}
top_TF_rf <- c("ABI3VP1_tnt.REM16_col_a_m1", "ARF_tnt.ARF16_col_b_m1", "E2FDP_tnt.DEL2_col_a_m1",	"E2FDP_tnt.DEL1_colamp_a_m1", "E2FDP_tnt.DEL2_colamp_a_m1", "E2FDP_tnt.DEL1_col_a_m1")

Clust %>% 
   filter(motif == "ABI3VP1_tnt.REM16_col_a_m1" | motif ==  "ARF_tnt.ARF16_col_b_m1" | motif ==  "E2FDP_tnt.DEL2_col_a_m1" |
            motif == "E2FDP_tnt.DEL1_colamp_a_m1" | motif == "E2FDP_tnt.DEL2_colamp_a_m1" | motif ==  "E2FDP_tnt.DEL1_col_a_m1")
          

top_TF_rf_hclust <- data_matrix %>% 
  filter(motif_1 == "ABI3VP1_tnt.REM16_col_a_m1" | motif_1 ==  "ARF_tnt.ARF16_col_b_m1" | motif_1 ==  "E2FDP_tnt.DEL2_col_a_m1" |
            motif_1 == "E2FDP_tnt.DEL1_colamp_a_m1" | motif_1 == "E2FDP_tnt.DEL2_colamp_a_m1" | motif_1 ==  "E2FDP_tnt.DEL1_col_a_m1") %>% 
  select(motif_1, "ABI3VP1_tnt.REM16_col_a_m1", "ARF_tnt.ARF16_col_b_m1", "E2FDP_tnt.DEL2_col_a_m1",	"E2FDP_tnt.DEL1_colamp_a_m1", "E2FDP_tnt.DEL2_colamp_a_m1", "E2FDP_tnt.DEL1_col_a_m1") 

test <- data_matrix_cluster_reordered %>% 
  filter(cluster_rows == 2 | cluster_rows == 3) %>% 
  select(motif_1, names(hclust_columns$Cluster_columns[hclust_columns$Cluster_columns == 1]))




```

ABI3VP1_tnt.REM16_col_a_m1,		
ARF_tnt.ARF16_col_b_m1,			
E2FDP_tnt.DEL2_col_a_m1,			
E2FDP_tnt.DEL1_colamp_a_m1,			
E2FDP_tnt.DEL2_colamp_a_m1,			
E2FDP_tnt.DEL1_col_a_m1
